requirements.txt
# Core data processing
datasets>=2.14.0
huggingface_hub>=0.17.0
pandas>=2.0.0
numpy>=1.24.0

# Code analysis tools
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-typescript>=0.20.0

# Static analysis and metrics
radon>=6.0.0
lizard>=1.17.0
gitpython>=3.1.0

# LLM APIs and evaluation
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
lighteval>=0.3.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
pyyaml>=6.0
jsonlines>=3.1.0
pathspec>=0.11.0

# Async processing
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Claude API access via Bearer Token (no additional dependencies needed)

# Testing and validation
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Retrieval-Augmented Generation (RAG)
sentence-transformers>=2.2.0

# Token counting (optional, for accurate OpenAI token counting)
tiktoken>=0.5.0

# Hugging Face models support
transformers>=4.30.0
torch>=2.0.0